+++ src/Constants/Country.php
@@ -2,6 +2,13 @@
 
 namespace PlacetoPay\PaymentMethod\Constants;
 
+use PlacetoPay\PaymentMethod\Countries\BelizeCountryConfig;
+use PlacetoPay\PaymentMethod\Countries\ChileCountryConfig;
+use PlacetoPay\PaymentMethod\Countries\CountryConfig;
+use PlacetoPay\PaymentMethod\Countries\EcuadorCountryConfig;
+use PlacetoPay\PaymentMethod\Countries\HondurasCountryConfig;
+use PlacetoPay\PaymentMethod\Countries\UruguayCountryConfig;
+
 interface Country
 {
     const BZ = 'BZ';
@@ -19,4 +26,15 @@ interface Country
     const PA = 'PA';
 
     const PR = 'PR';
+
+    const UY = 'UY';
+
+    public const COUNTRIES_CONFIG = [
+        EcuadorCountryConfig::class,
+        ChileCountryConfig::class,
+        HondurasCountryConfig::class,
+        BelizeCountryConfig::class,
+        UruguayCountryConfig::class,
+        CountryConfig::class
+    ];
 }