+++ src/Countries/CountryConfigInterface.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace PlacetoPay\PaymentMethod\Countries;
+
+interface CountryConfigInterface
+{
+    public static function resolve(string $countryCode): bool;
+    public static function getEndpoints(): array;
+}